Components: Extract a reusable PostFeaturedImage component
